backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
aiofiles
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
selenium
webdriver-manager
beautifulsoup4
gunicorn
